+++ pytket/docs/changelog.rst
@@ -15,7 +15,7 @@ Fixes:
 
 * Allow barriers when dagger or transpose a circuit.
 * Keep blank classical wires when running `FlattenRelabelRegistersPass`
-
+* Handle Clifford-angle ``NPhasedX`` gates in Clifford resynthesis.
 
 1.28.0 (May 2024)
 -----------------